refactor(codex): 拆分 thread listing 与 MCP 配置模块
- 将 unified thread listing/session merge 逻辑下沉到 thread_listing.rs\n- 将 global MCP config 解析逻辑下沉到 mcp_config.rs\n- 保持 codex command 暴露路径与现有能力不变，压缩 mod.rs 体量
